Clean up a few debug warning messages.

+++ plugins/rdioscanner_uploader/rdioscanner_uploader.cc
@@ -91,7 +91,7 @@ public:
     boost::filesystem::path audioName = audioPath.filename();
 
     if (call_info.transmission_source_list.size() != 0) {
-      for (int i = 0; i < call_info.transmission_source_list.size(); i++) {
+      for (unsigned long i = 0; i < call_info.transmission_source_list.size(); i++) {
         source_list << "{ \"pos\": " << std::setprecision(2) << call_info.transmission_source_list[i].position << ", \"src\": " << std::setprecision(0) << call_info.transmission_source_list[i].source << " }";
 
         if (i < (call_info.transmission_source_list.size() - 1)) {
@@ -105,7 +105,7 @@ public:
     }
 
     if (call_info.patched_talkgroups.size()>1){
-      for (int i = 0; i < call_info.patched_talkgroups.size(); i++) {
+      for (unsigned long i = 0; i < call_info.patched_talkgroups.size(); i++) {
         if (i!=0) { 
           patch_list << ",";
         }
